Find rows and columns in which all values are NA's

Find rows and columns in which all values are NA's

all_na_rows(x)

all_na_columns(x)

Arguments

x
A data frame or a matrix.

Value

A logical vector with rows/columns in which all values are NA indicated as TRUE.

Examples

 x <- matrix(c(NA, 1, 2, NA, 3, 4, NA, NA, NA), nrow = 3)
 x

#>      [,1] [,2] [,3]
#> [1,]   NA   NA   NA
#> [2,]    1    3   NA
#> [3,]    2    4   NA

 all_na_rows(x)

#> [1]  TRUE FALSE FALSE

 all_na_columns(x)

#> [1] FALSE FALSE  TRUE