[!] Clear variables from workspace (current environment)

Remove items from the workspace, i.e. the current environment, and free up system memory. Explanation in section "Details".

clear(..., list = NULL, except = NULL, all.names = FALSE,
  envir = parent.frame())

clear_all(..., list = NULL, except = NULL, all.names = TRUE,
  envir = parent.frame())

clear_hidden_only(..., list = NULL, except = NULL, envir = parent.frame())

clear_except(..., list = NULL, all.names = FALSE, envir = parent.frame())

clear_class(clrClass = NULL, exceptVar = NULL, all.names = FALSE,
  envir = parent.frame())

clear_except_class(exceptClass = NULL, all.names = FALSE,
  envir = parent.frame())

clear_fun(all.names = TRUE, envir = parent.frame())

Arguments

...
The objects as names (unquoted) or character strings (quoted).

list
A character vector naming objects. Used instead of `...`. If list is not NULL, dots `...` are ignored.

except, exceptVar
Names of variables (as a character vector) to be kept (to be cleared). Default is NULL.

all.names
a logical value. If TRUE (default only in clear_all and clear_fun), all objects are cleared. If FALSE (default in clear), objects which names begin with a . are omitted.

envir
The environment in which function takes action. Default is the global environment .GlobalEnv.

clrClass
Names of classes (as a character vector). Objects of indicated classes to be removed.

exceptClass
Names of classes (as a character vector) that have not to be cleared. Default is NULL.

Details

clear removes either listed or, if none is listed, all* not hidden (that name does not begin with dot(.)) variables. Default is to clear all except hidden objects. clear_all the same as clear just aditionally clears hidden variables (that name begins with dot (.)). clear_except clears variables except listed ones. clear_class removes objects of indicated class(es), except those which names provided in list exceptVar. clear_except_class keeps objects of indicated class(es), others are cleared. clear_fun removes all (including hidden) items that are functions.

These functions are wrappers and extensions for rm.

See also

Remove objects from a specified environment rm, list objects ls.

Other spMisc utilities: bru, clc, fCap, isFALSE, list_AddRm, make.filenames, open_wd, printDuration, regexp2df, st01

Examples

clear()

  A <- 5
  B <- "s"
 D1 <- "string2"
 D2 <- "string3"
  L <- list(A,B)
FUN <- function(x) x
`%in%` <- `%in%`

ls()

#> [1] "%in%" "A"    "B"    "D1"   "D2"   "FUN"  "L"   

clear_class("numeric")
ls()

#> [1] "%in%" "B"    "D1"   "D2"   "FUN"  "L"   

clear_fun()
ls()

#> [1] "B"  "D1" "D2" "L" 

clear(except = c("D1", "D2"))
ls()

#> [1] "D1" "D2"

clear(except = c("B"))
ls()

#> character(0)

clear()
ls()

#> character(0)

numeric_1 <- 5
numeric_2 <- 5
numeric_3 <- 5
My_list <- list("a","A")
My_string <- "ABC"
ls()

#> [1] "My_list"   "My_string" "numeric_1" "numeric_2" "numeric_3"

clear_except_class(c("numeric", "list"))
ls()

#> [1] "My_list"   "numeric_1" "numeric_2" "numeric_3"

clear_class("numeric", exceptVar = "numeric_1")
ls()

#> [1] "My_list"   "numeric_1"