Add code to HTML head that enables display of favicons

Function adds a line of code to the last line of head of all .html and .htm files in the indicated main directory and its subdirectories. NOTE: function doos not check if the favicon exists in the directory.

html_enable_favicons(main_dir = "./docs", favicon_path = "favicon.png")

Arguments

main_dir
(string(1)) root directory of your website.

favicon_path
(string(1)) relative path to favicon (from the main_dir).

Value

A vector with messages if procedure suceeded

See also

html_add_line_to_head

Examples

# html_enable_favicons("./docs", "favicon.ico")