[+] List all functions in a package

List all functions in a package and return it as a data frame.

list_functions(Package = "spMisc")

Arguments

Package
A name of a package. Default Package = "spMisc".

Value

A data frame with function names in a package.

Examples

list_functions()

#>       Functions in spMisc (0.0.22)
#> 1                            %!in%
#> 2                            %.+.%
#> 3                             %++%
#> 4                        %if.NULL%
#> 5                        %if_null%
#> 6                %if_null_or_len0%
#> 7                          %NOTin%
#> 8                   all_na_columns
#> 9                      all_na_rows
#> 10                             bru
#> 11                           bru_n
#> 12                            bru0
#> 13                       class_add
#> 14                             clc
#> 15                           clear
#> 16                       clear_all
#> 17                clear_all_except
#> 18          clear_all_except_class
#> 19                     clear_class
#> 20                    clear_except
#> 21                       clear_fun
#> 22               clear_hidden_only
#> 23                    corr_vec2mat
#> 24                        Duration
#> 25                           eval_
#> 26                       eval_glue
#> 27                            fCap
#> 28                       file2text
#> 29                 fixed_to_regexp
#> 30                   flattr_header
#> 31                        fullfile
#> 32                    getVarValues
#> 33           html_add_line_to_head
#> 34            html_enable_favicons
#> 35                         ifFALSE
#> 36                       indMatrix
#> 37                         isFALSE
#> 38                 knit_output_dir
#> 39                      list_AddRm
#> 40                  list_functions
#> 41                  make.filenames
#> 42                      na.replace
#> 43                      nDiag2nTri
#> 44                      nested_env
#> 45                      nTri2nDiag
#> 46                         open_wd
#> 47 pkgdown_find_missing_references
#> 48        print.missing_references
#> 49                   printDuration
#> 50                rbind_df_in_list
#> 51              regcapturedmatches
#> 52                       regexp2df
#> 53                    remove_names
#> 54                        rep_cols
#> 55                         rep_mat
#> 56                        rep_rows
#> 57                   revalueMatrix
#> 58                         rnw2rmd
#> 59                     simCorrVars
#> 60                     sort_colMax
#> 61                    sort_colSums
#> 62                 sort_descOnDiag
#> 63                     sort_rowMax
#> 64                    sort_rowSums
#> 65                            st01
#> 66                        stMinMax
#> 67                    time_elapsed
#> 68                   vignette_stub
#> 69                     visNetwork_
#> 70                        which.in
#> 71                    which.in.col
#> 72                   which.in.diag
#> 73                which.in.offdiag
#> 74                    which.in.row
#> 75                 which.in.trilow
#> 76                 which.in.triupp
#> 77                   which.max.all
#> 78                which.max.perCol
#> 79                which.max.perRow
#> 80                   which.min.all

list_functions(Package = "magrittr")

#>    Functions in magrittr (1.5)
#> 1                          %$%
#> 2                         %<>%
#> 3                          %>%
#> 4                         %T>%
#> 5                       [.fseq
#> 6                      [[.fseq
#> 7                          add
#> 8                          and
#> 9                   debug_fseq
#> 10                  debug_pipe
#> 11                   divide_by
#> 12               divide_by_int
#> 13                      equals
#> 14                     extract
#> 15                    extract2
#> 16                     freduce
#> 17                   functions
#> 18                       inset
#> 19                      inset2
#> 20            is_compound_pipe
#> 21                   is_dollar
#> 22                    is_first
#> 23                 is_function
#> 24                  is_funexpr
#> 25             is_greater_than
#> 26                       is_in
#> 27                is_less_than
#> 28            is_parenthesized
#> 29                     is_pipe
#> 30              is_placeholder
#> 31                      is_tee
#> 32      is_weakly_greater_than
#> 33         is_weakly_less_than
#> 34                         mod
#> 35                 multiply_by
#> 36          multiply_by_matrix
#> 37                   n'est pas
#> 38                         not
#> 39                          or
#> 40                        pipe
#> 41               prepare_first
#> 42            prepare_function
#> 43                  print.fseq
#> 44              raise_to_power
#> 45                set_colnames
#> 46                   set_names
#> 47                set_rownames
#> 48                 split_chain
#> 49                    subtract
#> 50                undebug_fseq
#> 51                  use_series
#> 52               wrap_function