[+!] Evaluate how much time passed

Evaluate and print difference between moment of time at which time interval began (usually captured by function Sys.time some time ago) and the present (i.e., moment when function printDuration is called).

printDuration(
  Start,
  Message = "Analysis completed in",
  returnString = FALSE,
  End = Sys.time()
)

Duration(
  Start,
  Message = "Analysis completed in",
  print = FALSE,
  End = Sys.time()
)

time_elapsed(start = stop("'start' is missing"), end = Sys.time())

Arguments

Start, start
Moment of time which is treated as the beggining (object of class POSIXct).

Message
Message before time stamp that describes it. Default is "Duration of analysis:"

returnString
If TRUE, returns result as a string. If FALSE (default), function pander prints the result.

End, end
Moment of time which is treated as the end (object of class POSIXct).

print
(logical) should the result be printed using package pander?

Value

Text indicating how much time has passed (either printed in console or as a string).

See also

Examples

Start <-  Sys.time()

Start

#> [1] "2020-03-17 19:07:04 EET"

## [1] "2016-02-12 16:15:09 UTC"

class(Start)

#> [1] "POSIXct" "POSIXt" 

## [1] "POSIXct" "POSIXt"

printDuration(Start)

#> Analysis completed in 0 secs

## Duration of analysis: 23.3 secs

printDuration(Start,"From start till now")

#> From start till now 0 secs

## From start till now 39.2 secs

my_duration <- printDuration(Start)

#> Analysis completed in 0 secs

my_duration

#> NULL

## NULL

my_duration <- printDuration(Start, returnString = TRUE)
my_duration

#> [1] "Analysis completed in 0 secs"

## [1] "Analysis completed in 2.4 mins"