[!+] Capture information to a dataframe by regular expressions

Capture information in substrings of text that match named and unnamed tokens of regular expressions and convert the result to a data frame.

regexp2df(text, pattern, ignore.case = FALSE, perl = TRUE,
  stringsAsFactors = default.stringsAsFactors(), ...)

Arguments

text
The text to be parsed: a character vector where matches are sought, or an object which can be coerced by as.character to a character vector.

pattern
Perl-like regular expression.

ignore.case
if FALSE, the pattern matching is case sensitive and if TRUE, case is ignored during matching.

perl
logical. Should Perl-compatible regexps be used?

stringsAsFactors
logical, passed to as.data.frame.

...
Other arguments to be passed to gregexpr.

Value

A data frame with parsed information.

Note

Call to function gregexpr with parameter perl = TRUE is used.

Contribution

In this function ideas from this answer on github.com were used.

See also

More about regular expressions used in R: regex Website handy for creating and testing Perl-like regular expressions (library pcre, version 1, not 2) https://regex101.com/r/dS3iP1/1#pcre

Functions gregexpr, regcapturedmatches, operator from package magrittr %>%.

Other spMisc utilities: bru, clc, clear, fCap, isFALSE, list_AddRm, make.filenames, open_wd, printDuration, st01

Examples

text1 <- c("A_111  B_aaa",
              "A_222  B_bbb",
              "A_333  B_ccc",
              "A_444  B_ddd",
              "A_555  B_eee")

# Named tokens
pattern1_named_tokens <- 'A_(?<Part_A>.*)  B_(?<Part_B>.*)'

regexp2df(text1, pattern1_named_tokens)

#>   Part_A Part_B
#> 1    111    aaa
#> 2    222    bbb
#> 3    333    ccc
#> 4    444    ddd
#> 5    555    eee

##     Part_A Part_B
## 1    111    aaa
## 2    222    bbb
## 3    333    ccc
## 4    444    ddd
## 5    555    eee

# Unnamed tockens - groups inside brackets:
pattern1_unnamed_tokens <- 'A_(.*)  B_(.*)'

regexp2df(text1, pattern1_unnamed_tokens)

#>     X X.1
#> 1 111 aaa
#> 2 222 bbb
#> 3 333 ccc
#> 4 444 ddd
#> 5 555 eee

##       X     X.1
## 1    111    aaa
## 2    222    bbb
## 3    333    ccc
## 4    444    ddd
## 5    555    eee

#----------------------------------------------------------
# Wrong. There must be NO SPACES in token's name:

pattern2 <- 'A (?<Part A>.*)  B (?<Part B>.*)'
regexp2df(text1, pattern2)

#> Warning: PCRE pattern compilation error
#> 	'syntax error in subpattern name (missing terminator)'
#> 	at ' A>.*)  B (?<Part B>.*)'

#> Error in gregexpr(pattern, text, ignore.case = ignore.case, perl = perl,     ...): invalid regular expression 'A (?<Part A>.*)  B (?<Part B>.*)'

## Error ...

#----------------------------------------------------------
text3 <- c("sn555 ID_O20-5-684_N52_2_Subt2_01.",
              "sn555 ID_O20-5-984_S52_8_Subt10_11.")

pattern3 <- paste0('sn(?<serial_number>.*) ID_(?<ID>.*)_(?<Class>[NS])',
                   '(?<Sector>.*)_(?<Point>.*)_[Ss]ubt.*\\.');

regexp2df(text3, pattern3)

#>   serial_number        ID Class Sector Point
#> 1           555 O20-5-684     N     52     2
#> 2           555 O20-5-984     S     52     8

##   serial_number    ID       Class Sector Point
## 1      555      O20-5-684     N     52     2
## 2      555      O20-5-984     S     52     8

#----------------------------------------------------------
# List all .R files in your working directory:

regexp2df(dir(),'(?<R_file>.*\\.[rR]$)')

#>    c.NA..NA..NA..NA..NA..NA..NA..NA..NA..NA..NA..NA..NA..NA..NA..
#> 1                                                              NA
#> 2                                                              NA
#> 3                                                              NA
#> 4                                                              NA
#> 5                                                              NA
#> 6                                                              NA
#> 7                                                              NA
#> 8                                                              NA
#> 9                                                              NA
#> 10                                                             NA
#> 11                                                             NA
#> 12                                                             NA
#> 13                                                             NA
#> 14                                                             NA
#> 15                                                             NA
#> 16                                                             NA
#> 17                                                             NA
#> 18                                                             NA
#> 19                                                             NA
#> 20                                                             NA
#> 21                                                             NA
#> 22                                                             NA
#> 23                                                             NA
#> 24                                                             NA
#> 25                                                             NA
#> 26                                                             NA
#> 27                                                             NA
#> 28                                                             NA
#> 29                                                             NA
#> 30                                                             NA
#> 31                                                             NA
#> 32                                                             NA
#> 33                                                             NA
#> 34                                                             NA
#> 35                                                             NA
#> 36                                                             NA
#> 37                                                             NA
#> 38                                                             NA
#> 39                                                             NA
#> 40                                                             NA
#> 41                                                             NA
#> 42                                                             NA
#> 43                                                             NA
#> 44                                                             NA
#> 45                                                             NA
#> 46                                                             NA
#> 47                                                             NA
#> 48                                                             NA
#> 49                                                             NA

# Do the same by using chaining operator %>%:
library(dplyr)

dir() %>% regexp2df('(?<R_file>\\.*[rR]$)')

#>    c.NA..NA..NA..NA..NA..NA..NA..NA..NA..NA..NA..NA..NA..NA..NA..
#> 1                                                              NA
#> 2                                                              NA
#> 3                                                              NA
#> 4                                                              NA
#> 5                                                              NA
#> 6                                                              NA
#> 7                                                              NA
#> 8                                                              NA
#> 9                                                              NA
#> 10                                                             NA
#> 11                                                             NA
#> 12                                                             NA
#> 13                                                             NA
#> 14                                                             NA
#> 15                                                             NA
#> 16                                                             NA
#> 17                                                             NA
#> 18                                                             NA
#> 19                                                             NA
#> 20                                                             NA
#> 21                                                             NA
#> 22                                                             NA
#> 23                                                             NA
#> 24                                                             NA
#> 25                                                             NA
#> 26                                                             NA
#> 27                                                             NA
#> 28                                                             NA
#> 29                                                             NA
#> 30                                                             NA
#> 31                                                             NA
#> 32                                                             NA
#> 33                                                             NA
#> 34                                                             NA
#> 35                                                             NA
#> 36                                                             NA
#> 37                                                             NA
#> 38                                                             NA
#> 39                                                             NA
#> 40                                                             NA
#> 41                                                             NA
#> 42                                                             NA
#> 43                                                             NA
#> 44                                                             NA
#> 45                                                             NA
#> 46                                                             NA
#> 47                                                             NA
#> 48                                                             NA
#> 49                                                             NA

#----------------------------------------------------------
# Capture several types of files:

expr <- paste0('(?<R_file>.*\\.[rR]$)|',
               '(?<Rmd_file>.*\\.[rR]md$)|',
               '(?<CSV_file>.*\\.[cC][sS][vV]$)')
dir() %>% regexp2df(expr)

#>    c.NA..NA..NA..NA..NA..NA..NA..NA..NA..NA..NA..NA..NA..NA..NA..
#> 1                                                              NA
#> 2                                                              NA
#> 3                                                              NA
#> 4                                                              NA
#> 5                                                              NA
#> 6                                                              NA
#> 7                                                              NA
#> 8                                                              NA
#> 9                                                              NA
#> 10                                                             NA
#> 11                                                             NA
#> 12                                                             NA
#> 13                                                             NA
#> 14                                                             NA
#> 15                                                             NA
#> 16                                                             NA
#> 17                                                             NA
#> 18                                                             NA
#> 19                                                             NA
#> 20                                                             NA
#> 21                                                             NA
#> 22                                                             NA
#> 23                                                             NA
#> 24                                                             NA
#> 25                                                             NA
#> 26                                                             NA
#> 27                                                             NA
#> 28                                                             NA
#> 29                                                             NA
#> 30                                                             NA
#> 31                                                             NA
#> 32                                                             NA
#> 33                                                             NA
#> 34                                                             NA
#> 35                                                             NA
#> 36                                                             NA
#> 37                                                             NA
#> 38                                                             NA
#> 39                                                             NA
#> 40                                                             NA
#> 41                                                             NA
#> 42                                                             NA
#> 43                                                             NA
#> 44                                                             NA
#> 45                                                             NA
#> 46                                                             NA
#> 47                                                             NA
#> 48                                                             NA
#> 49                                                             NA
#>    c.NA..NA..NA..NA..NA..NA..NA..NA..NA..NA..NA..NA..NA..NA..NA...1
#> 1                                                                NA
#> 2                                                                NA
#> 3                                                                NA
#> 4                                                                NA
#> 5                                                                NA
#> 6                                                                NA
#> 7                                                                NA
#> 8                                                                NA
#> 9                                                                NA
#> 10                                                               NA
#> 11                                                               NA
#> 12                                                               NA
#> 13                                                               NA
#> 14                                                               NA
#> 15                                                               NA
#> 16                                                               NA
#> 17                                                               NA
#> 18                                                               NA
#> 19                                                               NA
#> 20                                                               NA
#> 21                                                               NA
#> 22                                                               NA
#> 23                                                               NA
#> 24                                                               NA
#> 25                                                               NA
#> 26                                                               NA
#> 27                                                               NA
#> 28                                                               NA
#> 29                                                               NA
#> 30                                                               NA
#> 31                                                               NA
#> 32                                                               NA
#> 33                                                               NA
#> 34                                                               NA
#> 35                                                               NA
#> 36                                                               NA
#> 37                                                               NA
#> 38                                                               NA
#> 39                                                               NA
#> 40                                                               NA
#> 41                                                               NA
#> 42                                                               NA
#> 43                                                               NA
#> 44                                                               NA
#> 45                                                               NA
#> 46                                                               NA
#> 47                                                               NA
#> 48                                                               NA
#> 49                                                               NA
#>    c.NA..NA..NA..NA..NA..NA..NA..NA..NA..NA..NA..NA..NA..NA..NA...2
#> 1                                                                NA
#> 2                                                                NA
#> 3                                                                NA
#> 4                                                                NA
#> 5                                                                NA
#> 6                                                                NA
#> 7                                                                NA
#> 8                                                                NA
#> 9                                                                NA
#> 10                                                               NA
#> 11                                                               NA
#> 12                                                               NA
#> 13                                                               NA
#> 14                                                               NA
#> 15                                                               NA
#> 16                                                               NA
#> 17                                                               NA
#> 18                                                               NA
#> 19                                                               NA
#> 20                                                               NA
#> 21                                                               NA
#> 22                                                               NA
#> 23                                                               NA
#> 24                                                               NA
#> 25                                                               NA
#> 26                                                               NA
#> 27                                                               NA
#> 28                                                               NA
#> 29                                                               NA
#> 30                                                               NA
#> 31                                                               NA
#> 32                                                               NA
#> 33                                                               NA
#> 34                                                               NA
#> 35                                                               NA
#> 36                                                               NA
#> 37                                                               NA
#> 38                                                               NA
#> 39                                                               NA
#> 40                                                               NA
#> 41                                                               NA
#> 42                                                               NA
#> 43                                                               NA
#> 44                                                               NA
#> 45                                                               NA
#> 46                                                               NA
#> 47                                                               NA
#> 48                                                               NA
#> 49                                                               NA