[+] Simulate correlated continuous data from standard normal distribution

Simulate correlated continuaous data from standard normal distribution by using Choleski factorization of correlation matrix.

simCorrVars(corrMat, N = 100)

Arguments

corrMat
A symmetricmatrix of correlation coefficients. Determinant of corrMat must be positive. If corrMat is a vector, it is passed to function corr_vec2mat to construct a correlation matrix. Default corrMat = corr_vec2mat(c(.8,.2,.7)).

N
A number of observations. Default is 100.

Source

http://www.r-bloggers.com/simulating-random-multivariate-correlated-data-continuous-variables/

Value

A dataframe with correlated variables.

Details

Number of variables is determined by the size of correlation matrix corrMat.

See also

corr_vec2mat, mvrnorm, chol

Other simmulation functions in spMisc: corr_vec2mat

Examples

# ------------------------------------------------------------
simCorrVars()

#>               V1          V2           V3
#> 1   -1.400043517 -0.96684458 -0.994171492
#> 2   -0.005571287  0.36847460  1.003061094
#> 3   -1.821817661 -1.60584931 -0.681534405
#> 4   -0.282705449 -0.55838399 -0.311266838
#> 5    2.065024895  0.67342628 -0.856423379
#> 6   -1.863011492 -1.80361670 -0.862786194
#> 7    0.542996343 -0.11404782 -0.532752649
#> 8    0.362951256 -0.49236512 -0.815759399
#> 9    1.888504929  1.45233688 -0.072451730
#> 10  -0.015950311 -0.50883362 -1.333049987
#> 11   0.935363190  0.85418372  0.440291362
#> 12   1.623548883  1.36606196  0.373647231
#> 13  -1.910087468 -1.69561232 -0.754728116
#> 14   1.067307879  0.89586721  0.028961542
#> 15  -0.049964899 -0.19086199 -0.064058897
#> 16   2.755417575  2.23225289  0.816707518
#> 17   0.118194874 -1.05247640 -1.363024808
#> 18  -0.243236740 -0.31824171 -0.226698374
#> 19   0.029560754  0.35354513 -0.380003958
#> 20   2.682557184  1.92931299  0.294044633
#> 21   1.074345882  0.46042376  0.047721669
#> 22  -0.245896412 -0.90325511 -1.486931579
#> 23   1.065057320  0.93104824  0.520760159
#> 24  -1.699450568 -2.24200224 -1.553501834
#> 25   1.337320413  1.21187410  0.991063570
#> 26   0.523909788  0.78317666  0.608277297
#> 27   0.172181715  0.08354900  0.698436756
#> 28   1.298392759  1.48798897  1.149015449
#> 29  -0.548257264  0.22771512 -0.121922786
#> 30  -0.155693776  0.13577885  0.211433026
#> 31   0.424187575  0.97721126  1.447793962
#> 32  -0.038102895  0.26120704  1.077818098
#> 33  -0.354361164  0.28431980  1.290845520
#> 34  -0.296640025 -0.46964016 -0.712016982
#> 35  -1.056736867 -1.32271435 -1.607537208
#> 36  -0.690537897 -0.88755551 -0.848644187
#> 37   0.227127133  0.76877466  0.845134952
#> 38  -1.399410460 -0.96440600 -0.218306725
#> 39   0.568599861  1.73099016  2.192432352
#> 40  -1.684281532 -1.19778416  0.303113767
#> 41   2.039369263  1.90116768  1.351429515
#> 42   0.426566547  0.40580363  0.190773615
#> 43   0.603611011  0.32529846 -0.320259113
#> 44   0.192149422 -0.53400026 -0.665423906
#> 45   0.081719629 -0.71769450 -1.524224250
#> 46   0.454341594 -0.14964823 -0.789927972
#> 47   0.894961626  0.75635196  0.176562046
#> 48  -0.827310169  0.46405524  1.820234038
#> 49   0.979956696  1.57703395  0.951932092
#> 50   0.514599819 -0.49378005 -0.661641666
#> 51   0.429147371  0.41657995 -0.245027760
#> 52  -0.558015129  0.18511102  1.098147404
#> 53   0.038499547 -0.18302907 -0.009848942
#> 54   0.804411616 -0.49650720 -1.186744722
#> 55  -0.309051503 -0.08940120 -0.518543535
#> 56  -0.788258845 -1.31042008 -1.036529237
#> 57  -0.285887914  0.08189115  0.368868281
#> 58  -0.974069593 -0.01685229  1.320932484
#> 59   0.768721370  1.23653556  0.902546293
#> 60  -1.275334875 -1.20364030  0.326489057
#> 61  -1.041668381 -1.52124901 -1.889056616
#> 62   1.525938655  1.55326223  1.575882981
#> 63  -0.154120740  1.41534841  2.688453856
#> 64   1.142694878  1.58845921  1.086235918
#> 65  -0.823261151 -1.00593970 -0.002535704
#> 66   0.132992146  0.33229331  0.806467403
#> 67   1.241263075  1.36026503  0.632836273
#> 68   1.360461327  1.04585460  0.102915909
#> 69  -2.446680029 -1.91805204 -0.855848697
#> 70  -0.633178176 -1.74473521 -0.957997677
#> 71  -1.153398386 -1.12710143 -0.232696850
#> 72  -1.270513110 -0.69112556  0.262913163
#> 73   0.558514422  0.69605538 -0.076904638
#> 74   0.941206122  0.54960337 -0.146070840
#> 75   0.040204392  0.10674415 -0.266779426
#> 76   1.233390065  1.19096674  0.369983168
#> 77   0.708753061 -0.35037278 -1.142357790
#> 78  -1.312814246 -0.60203424 -0.195397908
#> 79   0.071053360 -0.32687817 -0.888713523
#> 80   0.675244698  1.23222123  0.519906676
#> 81  -0.902814949  0.06832826  1.431408994
#> 82   1.203767839  0.10425181 -0.511791058
#> 83   0.003125940 -0.04423134  0.101491463
#> 84   0.128922896 -0.39499024 -0.916448950
#> 85  -1.193641182 -1.40594694 -0.351434276
#> 86  -0.828603533 -0.48901815 -0.090847607
#> 87  -0.604829354  0.39220262  1.251103856
#> 88  -1.433321100 -1.15283887 -0.378135870
#> 89  -0.906340179 -1.98636363 -1.339909912
#> 90  -0.968125837 -0.83606249 -0.193031957
#> 91   0.060898893 -1.25782650 -1.993285684
#> 92   0.112294787  0.09456755  0.756803610
#> 93   2.158756554  2.15283396  1.367545766
#> 94  -0.308211421  0.36063197  0.493212227
#> 95   0.563380077  0.64419371  0.544922860
#> 96   1.129835153  0.33896928 -0.537013081
#> 97   1.415412293  0.90209001 -0.129700647
#> 98  -0.221744517 -0.78311285 -0.766740659
#> 99   1.604407328  0.37451115 -1.591064314
#> 100  0.876777327  1.07590131  1.555156112

simCorrVars(N = 10)

#>             V1         V2         V3
#> 1  -0.35612442 -0.9235781 -0.6120772
#> 2   1.18157557  1.0642957  0.2597917
#> 3   0.61615428  1.6774175  2.6298985
#> 4  -1.58862055 -1.5948503 -1.2565854
#> 5   0.55910599 -0.6443236 -1.3732499
#> 6   0.04213411  0.7415058  0.9706194
#> 7  -1.05633610 -0.7258026  0.2195828
#> 8   0.34391334  1.1616501  1.4264572
#> 9   2.12644453  0.8154375 -0.7453138
#> 10  1.39397780  1.3313492  0.8565522

# ------------------------------------------------------------
# Generate 2 correlated variables with 10 observations:
# A. Imput is a matrix

corrMat <- matrix(c(1, .5, .5, 1), 2)
print(corrMat)

#>      [,1] [,2]
#> [1,]  1.0  0.5
#> [2,]  0.5  1.0

simCorrVars(corrMat, N = 10)

#>             V1         V2
#> 1   1.05215542 -1.1882673
#> 2   1.20838561  0.4575919
#> 3   0.29502975  1.2441980
#> 4  -1.13534326 -1.5471930
#> 5   0.10999338  0.7936344
#> 6  -0.23433786  1.6899564
#> 7  -0.11091937 -1.2617006
#> 8  -1.14229077  0.9050891
#> 9  -0.08007363 -0.4187335
#> 10 -0.11921509  0.6214893

# B. Imput is a vector
simCorrVars(.5, N = 10)

#>            V1         V2
#> 1  -0.5789452 -0.4154160
#> 2   0.5264580  1.7645517
#> 3   1.4486572  2.0391224
#> 4  -0.3840073  1.3903332
#> 5  -0.5514917 -1.0255104
#> 6  -0.3438315  0.7485623
#> 7   0.8130582  1.9683915
#> 8  -0.1050687  0.7982952
#> 9  -1.7133026 -1.5772014
#> 10  1.1004919  0.3997133

simCorrVars(c(.1,.5,.8), N = 10)

#>              V1           V2         V3
#> 1   0.178812018 -0.677047326 -0.8465919
#> 2  -0.975423042 -0.434421671 -0.2515607
#> 3   0.405101183 -0.428051850 -0.2092390
#> 4   1.226682356  0.453943328  0.7163081
#> 5  -0.098550690  0.024736724  0.1415757
#> 6   0.020831228  0.009631871  0.4130476
#> 7  -0.684749941  0.267235272 -0.2637851
#> 8   0.934261130  1.924518151  1.5537854
#> 9   0.008510312  2.024844413  0.9654866
#> 10  1.158979182 -0.086292442  0.2651232

# ------------------------------------------------------------