Wrapper for function visNetwork

The result is the same as if function visNetwork was used.

visNetwork_(file, width = NULL, height = NULL, main = NULL,
  encoding = "UTF-8", ...)

file2text(file, encoding = "UTF-8")

Arguments

file
path to graphviz (.gv) file

width
: Width (optional, defaults to automatic sizing)

height
: Height (optional, defaults to automatic sizing)

main
: For add a title. Character or a named list.

"text" : Character. Title.

"style" : Optional. Character. HTML style of title. Default to 'font-family:Georgia, Times New Roman, Times, serif;font-weight:bold;font-size:20px;text-align:center;'.

encoding
encoding to be assumed for input strings. It is used to mark character strings as known to be in Latin-1 or UTF-8: it is not used to re-encode the input. To do the latter, specify the encoding as part of the connection con or via options(encoding=): see the examples. See also ‘Details’.

...
: Don't use.