[+] Get indices of certain matrix elements

Get indices of certain (e.g., diagonal) matrix elements. These elements can be in:

diag - diagonal;

offdiag - offdiagonal;

col - indicated columns;

row - indicated rows;

trilow - lower triangle (with or without diagonal);

triupp - upper triangle (with or without diagonal).

which.in(type, x, ...)

which.in.diag(x)

which.in.offdiag(x)

which.in.col(x, col)

which.in.row(x, row)

which.in.trilow(x, diag = FALSE)

which.in.triupp(x, diag = FALSE)

Arguments

type
Either diag, offdiag, col, row, trilow or triupp (just for which.in()).

x
A matrix or an object, which can be coerced to a marix.

...
Other appropriate parameters indicated below.

col
An index/Indices of columns that contain elements of interest.

row
An index/Indices of rows that contain elements of interest.

diag
logical. Should the diagonal be included?

Value

Indices of indicated elements of a matrix.

See also

Other matrix operations in spMisc: corr_vec2mat, indMatrix, nTri2nDiag, revalueMatrix, sort_matrix_elements, which.max.all, which.max.perRow

Examples

m1 <- matrix(NA, 5, 5)
m1

#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]   NA   NA   NA   NA   NA
#> [2,]   NA   NA   NA   NA   NA
#> [3,]   NA   NA   NA   NA   NA
#> [4,]   NA   NA   NA   NA   NA
#> [5,]   NA   NA   NA   NA   NA

 #>        [,1] [,2] [,3] [,4] [,5]
 #>  [1,]   NA   NA   NA   NA   NA
 #>  [2,]   NA   NA   NA   NA   NA
 #>  [3,]   NA   NA   NA   NA   NA
 #>  [4,]   NA   NA   NA   NA   NA
 #>  [5,]   NA   NA   NA   NA   NA

which.in.diag(m1)

#> [1]  1  7 13 19 25

 #>  [1]  1  7 13 19 25

m2 <- matrix(NA, 2, 5)
which.in.diag(m2)

#> [1] 1 4

 #>  [1]  1  4

#================================

 which.in(diag,    m1)

#> [1]  1  7 13 19 25

 which.in(offdiag, m1)

#>  [1]  2  3  4  5  6  8  9 10 11 12 14 15 16 17 18 20 21 22 23 24

 which.in(col, m1, col = 2)

#> [1]  6  7  8  9 10

 which.in(row, m1, row = 2)

#> [1]  2  7 12 17 22

 which.in(trilow, m1)

#>  [1]  2  3  4  5  8  9 10 14 15 20

 which.in(trilow, m1, diag = TRUE)

#>  [1]  1  2  3  4  5  7  8  9 10 13 14 15 19 20 25

 which.in(triupp, m1)

#>  [1]  6 11 12 16 17 18 21 22 23 24