[+] List all functions in a package

List all functions in a package and return it as a data frame.

list_functions(Package = "spMisc")

Arguments

Package
A name of a package. Default Package = "spMisc".

Value

A data frame with function names in a package.

Examples

list_functions()

#>    Functions in spMisc (0.0.16)
#> 1                         %!in%
#> 2                         %.+.%
#> 3                          %++%
#> 4                     %if.NULL%
#> 5                     %if_null%
#> 6             %if_null_or_len0%
#> 7                       %NOTin%
#> 8                all_na_columns
#> 9                   all_na_rows
#> 10                          bru
#> 11                        bru_n
#> 12                         bru0
#> 13                    class_add
#> 14                          clc
#> 15                        clear
#> 16                    clear_all
#> 17                  clear_class
#> 18                 clear_except
#> 19           clear_except_class
#> 20                    clear_fun
#> 21            clear_hidden_only
#> 22                 corr_vec2mat
#> 23                     Duration
#> 24                        eval_
#> 25                         fCap
#> 26                    file2text
#> 27              fixed_to_regexp
#> 28                     fullfile
#> 29                 getVarValues
#> 30        html_add_line_to_head
#> 31         html_enable_favicons
#> 32                      ifFALSE
#> 33                    indMatrix
#> 34                      isFALSE
#> 35                   list_AddRm
#> 36               list_functions
#> 37               make.filenames
#> 38                   na.replace
#> 39                   nDiag2nTri
#> 40                   nTri2nDiag
#> 41                      open_wd
#> 42                printDuration
#> 43             rbind_df_in_list
#> 44           regcapturedmatches
#> 45                    regexp2df
#> 46                     rep_cols
#> 47                      rep_mat
#> 48                     rep_rows
#> 49                revalueMatrix
#> 50                  simCorrVars
#> 51                  sort_colMax
#> 52                 sort_colSums
#> 53              sort_descOnDiag
#> 54                  sort_rowMax
#> 55                 sort_rowSums
#> 56                         st01
#> 57                     stMinMax
#> 58                 time_elapsed
#> 59                  visNetwork_
#> 60                     which.in
#> 61                 which.in.col
#> 62                which.in.diag
#> 63             which.in.offdiag
#> 64                 which.in.row
#> 65              which.in.trilow
#> 66              which.in.triupp
#> 67                which.max.all
#> 68             which.max.perCol
#> 69             which.max.perRow
#> 70                which.min.all

list_functions(Package = "magrittr")

#>    Functions in magrittr (1.5)
#> 1                          %$%
#> 2                         %<>%
#> 3                          %>%
#> 4                         %T>%
#> 5                       [.fseq
#> 6                      [[.fseq
#> 7                          add
#> 8                          and
#> 9                   debug_fseq
#> 10                  debug_pipe
#> 11                   divide_by
#> 12               divide_by_int
#> 13                      equals
#> 14                     extract
#> 15                    extract2
#> 16                     freduce
#> 17                   functions
#> 18                       inset
#> 19                      inset2
#> 20            is_compound_pipe
#> 21                   is_dollar
#> 22                    is_first
#> 23                 is_function
#> 24                  is_funexpr
#> 25             is_greater_than
#> 26                       is_in
#> 27                is_less_than
#> 28            is_parenthesized
#> 29                     is_pipe
#> 30              is_placeholder
#> 31                      is_tee
#> 32      is_weakly_greater_than
#> 33         is_weakly_less_than
#> 34                         mod
#> 35                 multiply_by
#> 36          multiply_by_matrix
#> 37                   n'est pas
#> 38                         not
#> 39                          or
#> 40                        pipe
#> 41               prepare_first
#> 42            prepare_function
#> 43                  print.fseq
#> 44              raise_to_power
#> 45                set_colnames
#> 46                   set_names
#> 47                set_rownames
#> 48                 split_chain
#> 49                    subtract
#> 50                undebug_fseq
#> 51                  use_series
#> 52               wrap_function